pyproject.toml
[project]
name = "stylegen"
version = "0.1.0"
description = "Gemini image generation CLI with style reference support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai",
    "python-dotenv",
]

[project.scripts]
sgen = "sgen:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sgen.py"]

[tool.hatch.build.targets.sdist]
include = ["sgen.py"]
